moving the few public types outside of ML.Model
