+++ docs/samples/Microsoft.ML.Samples/Dynamic/FeatureContributionCalculationTransform.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using Microsoft.ML.Data;
 using Microsoft.ML.SamplesUtils;
+using Microsoft.ML.Trainers;
 
 namespace Microsoft.ML.Samples.Dynamic
 {
