+++ src/Microsoft.ML.Data/Dirty/ModelParametersBase.cs
@@ -5,7 +5,7 @@
 using System;
 using Microsoft.ML.Runtime;
 
-namespace Microsoft.ML.Model
+namespace Microsoft.ML.Trainers
 {
     /// <summary>
     /// A base class for predictors producing <typeparamref name="TOutput"/>.
